Fix styles, change structure & export of Pages,Card
