+++ src/components/Card/Card.module.scss
@@ -16,16 +16,39 @@
 }
 
 .product_card {
-  background-color: $white;
+  background-color: #FFFFFF;
+  border: 1px solid #E2E6E9;
   display: flex;
   flex-direction: column;
   align-items: center;
-  width: 350px;//test
-  border: 1px solid $elements;
-  padding: 32px; 
-  
+  width: 100%;
+  max-width: 272px;
+  height: 506px;
+  padding: 32px;
+  box-sizing: border-box;
+
+  @media (min-width: 640px) and (max-width: 767px) {
+    max-width: 288px;
+  }
+
+  @media (min-width: 768px) and (max-width: 1199px) {
+    max-width: 229px;
+  }
+
+  @media (max-width: 639px) {
+    max-width: 287px;
+    height: 440px;
+  }
+
   &__image {
+    width: 100%;
+    height: 180px;
+    object-fit: contain;
     margin-bottom: 8px;
+
+    @media (max-width: 639px) {
+      max-height: 140px;
+    }
   }
 
   &__name {
@@ -49,13 +72,28 @@
     color: $primary;
     margin-bottom: 8px;
     width: 100%;
+    display: flex;
+    align-items: center;
+    gap: 8px;
+  }
+
+  &__price_discount {
+    font-size: 22px;
+    font-family: 'Mont-Bold', sans-serif;
+    font-weight: 700;
+    line-height: 30px;
+    color: $primary;
   }
 
   &__fullPrice {
-    text-decoration: line-through;
-    color: $secondary;
+    font-size: 22px;
     font-family: 'Mont-Regular', sans-serif;
-    font-weight: 500;
+    font-weight: 600;
+    line-height: 28.12px;
+    text-decoration-line: line-through;
+    text-decoration-skip-ink: none;
+    color: $secondary;
+    margin-left: 8px;
   }
 
   &__separator {
@@ -65,12 +103,15 @@
   }
 
   &__features {
-    padding: 8px 0;
+    flex-grow: 0;
     display: flex;
     flex-direction: column;
+    justify-content: space-between;
     gap: 8px;
     width: 100%;
-    margin-bottom: 8px;
+    min-width: 77px;
+    overflow: hidden;
+    padding: 8px;
   }
 
   &__feature {
