+++ src/components/Catalog/Catalog.module.scss
@@ -1,0 +1,139 @@
+.catalog {
+  &__container {
+    padding: 1rem;
+    width: auto;
+    max-width: 1600px;
+    margin: 0 auto;
+    height: auto;
+
+  }
+
+  &__title {
+    font-size: 2rem;
+    font-weight: bold;
+    margin-bottom: 1.5rem;
+  }
+
+  &__filters {
+    display: flex;
+    flex-wrap: wrap;
+    gap: 1rem;
+    margin-bottom: 1.5rem;
+  }
+
+  &__select {
+    border: 1px solid #d1d5db;
+    border-radius: 0.375rem;
+    padding: 0.5rem 0.75rem;
+  }
+
+  &__grid {
+    display: grid;
+    grid-template-columns: repeat(1, 287px); 
+    gap: 40px 16px;
+    padding: 32px;
+    justify-items: center;
+    align-items: center;
+
+    @media (min-width: 640px) and (max-width: 767px) {
+      grid-template-columns: repeat(2, 288px);
+      gap: 40px 16px;
+
+    }
+
+    @media (min-width: 768px) and (max-width: 1199px) {
+      grid-template-columns: repeat(3, 229px);
+      gap: 40px 16px;
+    }
+
+    @media (min-width: 1200px) {
+      grid-template-columns: repeat(4, 272px);
+      gap: 40px 16px;
+    }
+  }
+
+  &__card {
+    border: none;
+    padding: 0;
+    background-color: transparent;
+    box-sizing: border-box;
+    width: 100%;
+    transition: box-shadow 0.2s, background-color 0.2s;
+
+    &:hover {
+      box-shadow: 0px 2px 15px 0px #0000001A;
+      background-color: #f9fafb;
+    }
+  }
+
+  &__image {
+    width: 100%;
+    height: 10rem;
+    object-fit: contain;
+    border-radius: 0.375rem;
+    margin-bottom: 1rem;
+  }
+
+  &__name {
+    font-size: 1.125rem;
+    font-weight: 500;
+  }
+
+  &__price {
+    color: #6b7280;
+  }
+
+  &__pagination {
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    margin-top: 1.5rem;
+    gap: 0.5rem;
+  }
+
+  &__button {
+    padding: 0.25rem 0.75rem;
+    border: 1px solid #e5e7eb;
+    border-radius: 0.375rem;
+    background-color: #fff;
+    color: #374151;
+    cursor: pointer;
+    transition: background-color 0.2s;
+
+    &:hover {
+      background-color: #f3f4f6;
+    }
+
+    &--active {
+      background-color: #3b82f6;
+      color: #fff;
+      cursor: not-allowed;
+    }
+  }
+
+  &__message {
+    color: #6b7280;
+    font-size: 1.125rem;
+    text-align: center;
+    margin-top: 2rem;
+  }
+
+  &__error {
+    color: #ef4444;
+    font-size: 1.125rem;
+    margin-bottom: 1rem;
+    text-align: center;
+  }
+
+  &__reload {
+    padding: 0.5rem 1rem;
+    background-color: #3b82f6;
+    color: #fff;
+    border-radius: 0.375rem;
+    transition: background-color 0.2s;
+
+    &:hover {
+      background-color: #2563eb;
+    }
+  }
+}
