+++ src/pages/AccessoriesPage/AccessoriesPage.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
-import {Catalog} from '../../components/catalog';
-import {getAccessories} from '../../components/api/apiE';
+import { Catalog } from '../../components/Catalog/Catalog';
+import { getAccessories } from '../../components/api/apiE';
 
 export const AccessoriesPage: React.FC = () => {
-	return <Catalog fetchProducts={getAccessories} title='Accessories' />;
+  return <Catalog fetchProducts={getAccessories} title='Accessories' />;
 };
