+++ src/pages/PhonesPage/PhonesPage.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
-import {Catalog} from '../../components/catalog';
-import {getPhones} from '../../components/api/apiE';
+import { Catalog } from '../../components/Catalog/Catalog';
+import { getPhones } from '../../components/api/apiE';
 
 export const PhonesPage: React.FC = () => {
-	return <Catalog fetchProducts={getPhones} title='Phones' />;
+  return <Catalog fetchProducts={getPhones} title='Phones' />;
 };
