+++ src/pages/TabletsPage/TabletsPage.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
-import {Catalog} from '../../components/catalog';
-import {getTablets} from '../../components/api/apiE';
+import { Catalog } from '../../components/Catalog/Catalog';
+import { getTablets } from '../../components/api/apiE';
 
 export const TabletsPage: React.FC = () => {
-	return <Catalog fetchProducts={getTablets} title='Tablets' />;
+  return <Catalog fetchProducts={getTablets} title='Tablets' />;
 };
