+++ src/types/Product.ts
@@ -1,15 +1,16 @@
 export interface Product {
-	priceRegular: number;
-	id: number;
-	category: string;
-	itemId: string;
-	name: string;
-	fullPrice: number;
-	price?: number;
-	screen: string;
-	capacity: string;
-	color: string;
-	ram: string;
-	year: number;
-	images: string;
+  priceRegular: number;
+  priceDiscount?: number;
+  id: number;
+  category: string;
+  itemId: string;
+  name: string;
+  fullPrice: number;
+  price?: number;
+  screen: string;
+  capacity: string;
+  color: string;
+  ram: string;
+  year: number;
+  images: string;
 }
