+++ src/types/ProductSpec.ts
@@ -1,26 +1,27 @@
 export interface ProductDescription {
-	title: string;
-	text: string[];
+  title: string;
+  text: string[];
 }
 
 export interface ProductSpec {
-	id: string;
-	category: string;
-	namespaceId: string;
-	name: string;
-	capacityAvailable: string[];
-	capacity: string;
-	priceRegular: number;
-	priceDiscount?: number;
-	colorsAvailable: string[];
-	color: string;
-	images: string[];
-	description: ProductDescription[];
-	screen: string;
-	resolution: string;
-	processor: string;
-	ram: string;
-	camera?: string;
-	zoom?: string;
-	cell?: string[];
+  id: string;
+  category: string;
+  namespaceId: string;
+  name: string;
+  capacityAvailable: string[];
+  capacity: string;
+  priceRegular: number;
+  priceDiscount?: number;
+  colorsAvailable: string[];
+  color: string;
+  images: string[];
+  description: ProductDescription[];
+  screen: string;
+  resolution: string;
+  processor: string;
+  ram: string;
+  camera?: string;
+  zoom?: string;
+  cell?: string[];
+  year?: number;
 }
